adding tests for Frame - moved tests from FrameFreqHelper into tests for Frame
