+++ test/terra/pulse_new/models/test_operator_models.py
@@ -15,7 +15,7 @@ import unittest
 import numpy as np
 from scipy.linalg import expm
 from qiskit.quantum_info.operators import Operator
-from qiskit.providers.aer.pulse_new.models.operator_models import FrameFreqHelper, OperatorModel
+from qiskit.providers.aer.pulse_new.models.operator_models import OperatorModel
 from qiskit.providers.aer.pulse_new.models.signals import Constant, Signal, VectorSignal
 
 
@@ -226,7 +226,7 @@ class TestOperatorModel(unittest.TestCase):
 
     def test_cutoff_freq(self):
         """Test cutoff_freq"""
-        
+
         # enter frame of drift
         self.basic_model.frame_operator = self.basic_model.drift
 
@@ -358,191 +358,3 @@ class TestOperatorModel(unittest.TestCase):
 
     def assertAlmostEqual(self, A, B, tol=1e-12):
         self.assertTrue(np.abs(A - B).max() < tol)
-
-
-class TestFrameFreqHelper(unittest.TestCase):
-
-    def setUp(self):
-        self.X = Operator.from_label('X')
-        self.Y = Operator.from_label('Y')
-        self.Z = Operator.from_label('Z')
-
-
-    def test_evaluate_no_cutoff(self):
-        """test evaluate with a non-diagonal frame and no cutoff freq."""
-
-        frame_op = -1j * np.pi * self.X
-        operators = [Operator(-1j * np.pi * self.Z), Operator(-1j * self.X / 2)]
-        carrier_freqs = np.array([0., 1.])
-
-        helper = FrameFreqHelper(operators, carrier_freqs, frame_op)
-
-        t = np.pi * 0.02
-        coeffs = np.array([1., 1.])
-        val = helper.evaluate(t, coeffs)
-        U = expm(frame_op.data * t)
-        U_adj = U.conj().transpose()
-        expected = (U_adj @ (-1j * np.pi * self.Z.data +
-                             1j * np.pi * self.X.data +
-                            -1j * np.cos(2 * np.pi * t) * self.X.data / 2) @ U)
-
-        self.assertAlmostEqual(val, expected)
-
-        # with complex envelope
-        t = np.pi * 0.02
-        coeffs = np.array([1., 1. + 2 * 1j])
-        val = helper.evaluate(t, coeffs)
-        U = expm(frame_op.data * t)
-        U_adj = U.conj().transpose()
-        expected = (U_adj @ (-1j * np.pi * self.Z.data +
-                             1j * np.pi * self.X.data +
-                            -1j * np.cos(2 * np.pi * t) * self.X.data / 2 +
-                            1j * 2 * np.sin(2 * np.pi * t) * self.X.data / 2) @ U)
-
-        self.assertAlmostEqual(val, expected)
-
-    def test_evaluate_diag_frame_no_cutoff(self):
-        """test evaluate with a diagonal frame and no cutoff freq."""
-
-        frame_op = -1j * np.pi * np.array([1., -1.])
-        operators = [Operator(-1j * np.pi * self.Z), Operator(-1j * self.X / 2)]
-        carrier_freqs = np.array([0., 1.])
-
-        helper = FrameFreqHelper(operators, carrier_freqs, frame_op)
-
-        t = np.pi * 0.02
-        coeffs = np.array([1., 1.])
-        val = helper.evaluate(t, coeffs)
-        U = np.diag(np.exp(frame_op * t))
-        U_adj = U.conj().transpose()
-        expected = -1j * np.cos(2 * np.pi * t) * U_adj @ self.X.data @ U / 2
-
-        self.assertAlmostEqual(val, expected)
-
-        # with complex envelope
-        t = np.pi * 0.02
-        coeffs = np.array([1., 1. + 2 * 1j])
-        val = helper.evaluate(t, coeffs)
-        U = np.diag(np.exp(frame_op * t))
-        U_adj = U.conj().transpose()
-        expected = -1j * (np.cos(2 * np.pi * t) * U_adj @ self.X.data @ U -
-                    2 * np.sin(2 * np.pi * t) * U_adj @ self.X.data @ U ) / 2
-
-        self.assertAlmostEqual(val, expected)
-
-    def test_evaluate_no_frame(self):
-        """Test FrameFreqHelper.evaluate with no frame or cutoff."""
-
-        operators = [self.X, self.Y, self.Z]
-        carrier_freqs = np.array([1., 2., 3.])
-
-        ffhelper = FrameFreqHelper(operators, carrier_freqs)
-
-        t = 0.123
-        coeffs = np.array([1., 1j, 1 + 1j])
-
-        out = ffhelper.evaluate(t, coeffs)
-        sig_vals = np.real(coeffs * np.exp(1j * 2 * np.pi * carrier_freqs * t))
-        ops_as_arrays = np.array([op.data for op in operators])
-        expected_out = np.tensordot(sig_vals, ops_as_arrays, axes=1)
-
-        self.assertAlmostEqual(out, expected_out)
-
-        t = 0.123 * np.pi
-        coeffs = np.array([4.131, 3.23, 2.1 + 3.1j])
-
-        out = ffhelper.evaluate(t, coeffs)
-        sig_vals = np.real(coeffs * np.exp(1j * 2 * np.pi * carrier_freqs * t))
-        ops_as_arrays = np.array([op.data for op in operators])
-        expected_out = np.tensordot(sig_vals, ops_as_arrays, axes=1)
-
-        self.assertAlmostEqual(out, expected_out)
-
-    def test_internal_helper_mats_no_cutoff(self):
-        """Test internal setup steps for helper matrices with no cutoff freq."""
-
-        # no cutoff with already diagonal frame
-        frame_op = -1j * np.pi * np.array([1., -1.])
-        operators = [self.X, self.X, self.X]
-        carrier_freqs = np.array([1., 2., 3.])
-
-        helper = FrameFreqHelper(operators, carrier_freqs, frame_op)
-
-        D_diff = -1j * np.pi * np.array([[0, -2.], [2., 0.]])
-        im_freqs = 1j * 2 * np.pi * carrier_freqs
-        S_expect = np.array([w + D_diff for w in im_freqs])
-        M_expect = None
-
-        self.assertTrue(helper._cutoff_array == M_expect)
-        self.assertAlmostEqual(helper._freq_array, S_expect)
-
-        # same test but with frame given as a 2d array
-        # in this case diagonalization will occur, causing eigenvalues to
-        # be sorted in ascending order
-        frame_op = -1j * np.pi * np.array([[-1., 0], [0, 1.]])
-        operators = [self.X, self.X, self.X]
-        carrier_freqs = np.array([1., 2., 3.])
-
-        helper = FrameFreqHelper(operators, carrier_freqs, frame_op)
-
-        D_diff = -1j * np.pi * np.array([[0, 2.], [-2., 0.]])
-        im_freqs = 1j * 2 * np.pi * carrier_freqs
-        S_expect = np.array([w + D_diff for w in im_freqs])
-        M_expect = None
-
-        self.assertTrue(helper._cutoff_array == M_expect)
-        self.assertAlmostEqual(helper._freq_array, S_expect)
-
-    def test_internal_helper_mats_with_cutoff(self):
-        """Test internal setup steps for helper matrices with cutoff freq."""
-
-        # cutoff test
-        frame_op = -1j * np.pi * np.array([1., -1.])
-        operators = [self.X, self.X, self.X]
-        carrier_freqs = np.array([1., 2., 3.])
-        cutoff_freq = 3.
-
-        helper = FrameFreqHelper(operators,
-                                 carrier_freqs,
-                                 frame_op,
-                                 cutoff_freq)
-
-        D_diff = -1j * np.pi * np.array([[0, -2.], [2., 0.]])
-        im_freqs = 1j * 2 * np.pi * carrier_freqs
-        S_expect = np.array([w + D_diff for w in im_freqs])
-        M_expect = np.array([[[1, 1],
-                              [1, 1]],
-                             [[1, 0],
-                              [1, 1]],
-                             [[0, 0],
-                              [1, 0]]
-                             ])
-
-        self.assertAlmostEqual(helper._cutoff_array, M_expect)
-        self.assertAlmostEqual(helper._freq_array, S_expect)
-
-        # same test with lower cutoff
-        cutoff_freq = 2.
-
-        helper = FrameFreqHelper(operators,
-                                 carrier_freqs,
-                                 frame_op,
-                                 cutoff_freq)
-
-        D_diff = -1j * np.pi * np.array([[0, -2.], [2., 0.]])
-        im_freqs = 1j * 2 * np.pi * carrier_freqs
-        S_expect = np.array([w + D_diff for w in im_freqs])
-        M_expect = np.array([[[1, 0],
-                              [1, 1]],
-                             [[0, 0],
-                              [1, 0]],
-                             [[0, 0],
-                              [0, 0]]
-                             ])
-
-        self.assertAlmostEqual(helper._cutoff_array, M_expect)
-        self.assertAlmostEqual(helper._freq_array, S_expect)
-
-    def assertAlmostEqual(self, A, B, tol=1e-15):
-        diff = np.abs(A - B).max()
-        self.assertTrue(diff < tol)
